File/folder copy (#41)
* File/folder copy

* PR fixes

+++ app/src/main/java/com/bytemedrive/file/root/FileViewModel.kt
@@ -16,6 +16,7 @@ import com.bytemedrive.folder.EventFolderDeleted
 import com.bytemedrive.folder.EventFolderStarAdded
 import com.bytemedrive.folder.EventFolderStarRemoved
 import com.bytemedrive.folder.Folder
+import com.bytemedrive.folder.FolderManager
 import com.bytemedrive.navigation.AppNavigator
 import com.bytemedrive.privacy.AesService
 import com.bytemedrive.store.AppState
@@ -30,6 +31,7 @@ class FileViewModel(
     private val fileRepository: FileRepository,
     private val eventPublisher: EventPublisher,
     private val appNavigator: AppNavigator,
+    private val folderManager: FolderManager
 ) : ViewModel() {
 
     var files = MutableStateFlow(AppState.customer.value!!.files)
@@ -106,8 +108,15 @@ class FileViewModel(
 
     fun removeFile(id: UUID, onSuccess: () -> Unit) = viewModelScope.launch {
         AppState.customer.value?.wallet?.let { walletId ->
+            val file = files.value.find { it.id == id }
+            val physicalFileRemovable = files.value.none { it.chunkId == file?.chunkId }
+
             eventPublisher.publishEvent(EventFileDeleted(id))
-            fileRepository.remove(walletId, id)
+
+            if (physicalFileRemovable) {
+                fileRepository.remove(walletId, id)
+            }
+
             onSuccess()
         }
     }
@@ -115,11 +124,16 @@ class FileViewModel(
     fun removeFolder(id: UUID, onSuccess: () -> Unit) = viewModelScope.launch {
         AppState.customer.value?.wallet?.let { walletId ->
             folders.value.find { it.id == id }?.let { folder ->
-                findFilesToRemove(id, folders.value, files.value).forEach { file ->
+                findAllFilesRecursively(id, folders.value, files.value).forEach { file ->
+                    val physicalFileRemovable = files.value.none { it.chunkId == file.chunkId }
+
                     eventPublisher.publishEvent(EventFileDeleted(file.id))
-                    fileRepository.remove(walletId, file.id)
+
+                    if (physicalFileRemovable) {
+                        fileRepository.remove(walletId, file.id)
+                    }
                 }
-                (findFoldersToRemove(id, folders.value) + folder).forEach {
+                (folderManager.findAllFoldersRecursively(id, folders.value) + folder).forEach {
                     eventPublisher.publishEvent(EventFolderDeleted(it.id))
                 }
             }
@@ -146,37 +160,21 @@ class FileViewModel(
         onSuccess()
     }
 
-    private fun findFilesToRemove(folderId: UUID, allFolders: List<Folder>, allFiles: List<File>): List<File> {
-        val filesToRemove = allFiles.filter { it.folderId == folderId }
-        val subFolders = allFolders.filter { it.parent == folderId }
-
-        val filesInSubFolders = mutableListOf<File>()
-        for (subfolder in subFolders) {
-            filesInSubFolders.addAll(findFilesToRemove(subfolder.id, allFolders, allFiles))
-        }
-
-        return filesToRemove + filesInSubFolders
-    }
-
-    private fun findFoldersToRemove(folderId: UUID, allFolders: List<Folder>): List<Folder> {
-        val objectsInFolder = allFolders.filter { it.parent == folderId }
-
-        val objectsInSubFolders = mutableListOf<Folder>()
-        for (subfolder in objectsInFolder) {
-            objectsInSubFolders.addAll(findFoldersToRemove(subfolder.id, allFolders))
-        }
-
-        return objectsInFolder + objectsInSubFolders
-    }
-
     fun getFilesPages(): Flow<PagingData<Item>> =
         Pager(
             config = PagingConfig(pageSize = 20),
             pagingSourceFactory = { FilePagingSource(fileAndFolderList.value) }
         ).flow
 
-    fun prepareItemsToMove(ids: List<UUID>) {
+    fun useSelectionScreenToMoveItems(id: UUID) = useSelectionScreenToMoveItems(listOf(id))
+
+    fun useSelectionScreenToMoveItems(ids: List<UUID>) {
         action.value = Action(ids, Action.Type.MoveItems)
+        fileSelectionDialogOpened.value = true
+    }
+
+    fun useSelectionScreenToCopyItem(id: UUID) {
+        action.value = Action(listOf(id), Action.Type.CopyItem)
         fileSelectionDialogOpened.value = true
     }
 
@@ -198,6 +196,18 @@ class FileViewModel(
             }
         }
 
+    private fun findAllFilesRecursively(folderId: UUID, allFolders: List<Folder>, allFiles: List<File>): List<File> {
+        val filesToRemove = allFiles.filter { it.folderId == folderId }
+        val subFolders = allFolders.filter { it.parent == folderId }
+
+        val filesInSubFolders = mutableListOf<File>()
+        for (subfolder in subFolders) {
+            filesInSubFolders.addAll(findAllFilesRecursively(subfolder.id, allFolders, allFiles))
+        }
+
+        return filesToRemove + filesInSubFolders
+    }
+
     private suspend fun getThumbnails() {
         thumbnails.value = files.value.associate {
             it.id to it.thumbnails
@@ -214,6 +224,7 @@ class FileViewModel(
 
 data class Action(val ids: List<UUID>, val type: Type) {
     enum class Type(type: String) {
-        MoveItems("moveItems"),
+        CopyItem("copyItem"),
+        MoveItems("moveItems")
     }
 }
